backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# LangChain (支持最新版本)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0

# Vector Database (使用更新版本)
chromadb>=0.4.0

# Document Processing
pymupdf>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0

# Advanced RAG Dependencies
rank-bm25>=0.2.2        # BM25 keyword search
jieba>=0.42.1           # Chinese word segmentation for BM25
httpx>=0.27.0           # Async HTTP client for Rerank API

# Note: pydantic is auto-installed by langchain